leds: leds-atc: Keep BCL above LPM when CHG_LED is on
If charger goes to sleep mode, the CHG_LED will be turned off. Keeping
BCL above LPM when CHG_LED is on to keep the charger in standby mode to
avoid messing up the CHG_LED behavior.

This is a port to moto led drivers of 7bb9de3366e9014bb295b56bb.

Test: Blinking patterns are now working correctly in sleep mode

Change-Id: I43de4bd2f8e96954a5388724539702dd0662beaf
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>